Refactor imports to use the new bootstrap configuration module
