+++ integrations/standard_io/config.py
@@ -6,8 +6,8 @@ import logging
 from dataclasses import dataclass
 from typing import TYPE_CHECKING
 
-from cls.config import BaseSection
 from integrations.base.interface import IntegrationsConfig
+from libs.bootstrap.config import BaseSection
 
 if TYPE_CHECKING:
     from pathlib import Path  # noqa: F401
