+++ libs/global_value.py
@@ -3,11 +3,11 @@
 from typing import TYPE_CHECKING, Any, Callable, Union
 
 if TYPE_CHECKING:
-    from cls.config import AppConfig
     from integrations.discord.adapter import ServiceAdapter as discord_adapter
     from integrations.slack.adapter import ServiceAdapter as slack_adapter
     from integrations.standard_io.adapter import ServiceAdapter as std_adapter
     from integrations.web.adapter import ServiceAdapter as web_adapter
+    from libs.bootstrap.config import AppConfig
     from libs.types import Args, PlaceholderDict
 
 # --- グローバル変数 ---
