+++ tests/database/conftest.py
@@ -11,8 +11,8 @@ import pandas as pd
 import pytest
 
 import libs.global_value as g
-from cls.config import AppConfig
 from libs.bootstrap import configuration
+from libs.bootstrap.config import AppConfig
 from libs.data import initialization, lookup
 from libs.utils import dbutil
